split: install tests into separate files with live output
