--- serving/skills-cli/install-dist.test.ts
@@ -1,127 +1,0 @@
-import test from 'node:test';
-import assert from 'node:assert';
-import fs from 'node:fs';
-import path from 'node:path';
-import { execSync } from 'node:child_process';
-import { createIsolatedHome, cleanupIsolatedHome } from '../../harness/lib/agent-shared.ts';
-
-// Assuming assertSearchResults is available or we duplicate it
-function assertSearchResults(output: string) {
-    const results = JSON.parse(output);
-    assert.ok(Array.isArray(results), 'Output should be a JSON array');
-    assert.ok(results.length > 0, 'Search should find some results');
-}
-
-test('Claude Code loads plugin from local dist directory', { skip: !process.env.FULL }, async () => {
-    let homeDir = '';
-    try {
-        homeDir = createIsolatedHome('test-install-claude');
-        const distDir = path.resolve(import.meta.dirname, '../../dist/skills-cli');
-        
-        if (!fs.existsSync(distDir)) {
-            test.skip('dist/skills-cli not found, skipping');
-            return;
-        }
-
-        // Claude Code uses --plugin-dir to load local plugins
-        // We verify if we can run it and if it doesn't crash (or if it lists the plugin if we can find a way to list it)
-        // Since we don't want to run an interactive session, we just check if it parses the flag and runs a prompt.
-        // Filter for Anthropic environment variables to pass through securely in-memory
-        const anthropicEnv: Record<string, string> = {};
-        for (const [key, value] of Object.entries(process.env)) {
-            if (key.startsWith('ANTHROPIC_') && value) {
-                anthropicEnv[key] = value;
-            }
-        }
-
-        const cmd = `claude --plugin-dir ${distDir} -p "use the modern-web-use-cases skill and tell me best practices on implementing an address form" --dangerously-skip-permissions`;
-        const output = execSync(cmd, { 
-            encoding: 'utf8', 
-            timeout: 90000, // 90s
-            env: { ...process.env, HOME: homeDir, ...anthropicEnv } 
-        });
-        assert.ok(output, 'Claude should return some output');
-    } finally {
-        if (homeDir) {
-            cleanupIsolatedHome(homeDir);
-        }
-    }
-});
-
-test('Gemini CLI verifies extension install capability', { skip: !process.env.FULL }, async () => {
-    let homeDir = '';
-    try {
-        homeDir = createIsolatedHome('test-install-gemini');
-        const distDir = path.resolve(import.meta.dirname, '../../dist/skills-cli');
-        
-        const geminiBin = path.resolve(import.meta.dirname, '../../harness/node_modules/.bin/gemini');
-        if (!fs.existsSync(geminiBin)) {
-            test.skip('Gemini binary not found, skipping');
-            return;
-        }
-
-        // Verify if help works
-        const helpOut = execSync(`${geminiBin} extensions --help`, { encoding: 'utf8' });
-        assert.ok(helpOut.includes('install'), 'Gemini should have install command');
-
-        // We use 'yes |' to bypass the trust workspace prompt, and --consent for the security prompt
-        const installOutput = execSync(`yes | ${geminiBin} extensions install ${distDir} --consent`, { 
-            encoding: 'utf8', 
-            env: { ...process.env, HOME: homeDir },
-            stdio: 'pipe' 
-        });
-        // Loosen assertion to allow any output as long as it doesn't throw, or check for non-empty
-        assert.ok(installOutput !== undefined, 'Install should complete running');
-
-        // Verify functionality by running a prompt that triggers the skill
-        const promptCmd = `${geminiBin} -p "use the modern-web-use-cases skill and tell me best practices on implementing an address form" -o stream-json --yolo`;
-        const promptOutput = execSync(promptCmd, { 
-            encoding: 'utf8', 
-            timeout: 90000, // 90s
-            env: { ...process.env, HOME: homeDir },
-            stdio: 'pipe'
-        });
-        assert.ok(promptOutput, 'Gemini should return output for the skill prompt');
-    } finally {
-        if (homeDir) {
-            cleanupIsolatedHome(homeDir);
-        }
-    }
-});
-
-test('npx skills add from local path', { skip: !process.env.FULL }, async () => {
-    let homeDir = '';
-    try {
-        homeDir = createIsolatedHome('test-install-skills');
-        const distDir = path.resolve(import.meta.dirname, '../../dist/skills-cli');
-        
-        if (!fs.existsSync(distDir)) {
-            test.skip('dist/skills-cli not found, skipping');
-            return;
-        }
-
-        const cmd = `npx skills add -y -g ${distDir}`;
-        
-        try {
-            console.log(`Running: ${cmd}`);
-            const output = execSync(cmd, { 
-                encoding: 'utf8', 
-                stdio: 'pipe',
-                env: { ...process.env, HOME: homeDir, DISABLE_TELEMETRY: '1' }
-            });
-            console.log('npx skills add output:', output);
-            assert.ok(output.includes('Installed 1 skill') || output.includes('Installation complete'), 'Skills add should succeed');
-        } catch (e: any) {
-            console.log('npx skills add failed:', e.message);
-            if (e.message.includes('Invalid URL') || e.message.includes('Only GitHub repositories are supported')) {
-                console.log('npx skills add likely only supports GitHub repositories.');
-            } else {
-                throw e;
-            }
-        }
-    } finally {
-        if (homeDir) {
-            cleanupIsolatedHome(homeDir);
-        }
-    }
-});
